refactor: 첫 화면을 컴포넌트로 분리
